requirements.txt
inspect_ai
google-genai
openai
anthropic

# AWAB Data Generation dependencies
pydantic>=2.0.0
click>=8.0.0
pyyaml>=6.0.0
jinja2>=3.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
